Version 3.3.4

+++ Boxscore.php
@@ -7,6 +7,8 @@ $GameNumber = (integer)0;
 $GameYear = (integer)0;
 $GameHTML = (string)"";
 $YearH1 = (integer)0;
+$Playoff = (boolean)False;
+$AllStar = (boolean)False;
 
 If (file_exists($DatabaseFile) == false){
 	$LeagueName = $DatabaseNotFound;
@@ -16,23 +18,36 @@ If (file_exists($DatabaseFile) == false){
 	
 	$db = new SQLite3($DatabaseFile);
 
-	$Query = "Select Name, LeagueYear from LeagueGeneral";
+	$Query = "Select Name, LeagueYear, PlayOffStarted, OutputName from LeagueGeneral";
 	$LeagueGeneral = $db->querySingle($Query,true);		
 	$LeagueName = $LeagueGeneral['Name'];
 	$GameYear = $LeagueGeneral['LeagueYear'];
 	
-	$Query = "Select OutputGameHTMLToSQLiteDatabase from LeagueOutputOption";
+	$Query = "Select OutputGameHTMLToSQLiteDatabase, WebsiteURL from LeagueOutputOption";
 	$LeagueOutputOption = $db->querySingle($Query,true);	
 	
 	if(isset($_GET['Game'])){$GameNumber = filter_var($_GET['Game'], FILTER_SANITIZE_NUMBER_INT);} 
-	if(isset($_GET['Year'])){$GameYear = filter_var($_GET['Year'], FILTER_SANITIZE_NUMBER_INT);$YearH1=$GameYear;}
+	if(isset($_GET['Year'])){
+		$GameYear = filter_var($_GET['Year'], FILTER_SANITIZE_NUMBER_INT);$YearH1=$GameYear;
+		if(isset($_GET['Playoff'])){$Playoff=True;}
+	}else{
+		if($LeagueGeneral['PlayOffStarted'] == "True"){$Playoff=True;}
+	}
+	
 	If ($GameNumber > 0){
 		If ($LeagueOutputOption['OutputGameHTMLToSQLiteDatabase'] == "True"){
+			If($Playoff == True){$GameHTMLDatabaseFile = str_replace("-STHSGame","-PLF-STHSGame",$GameHTMLDatabaseFile);}
 			$GameDatabaseFile = str_replace("@-@",$GameYear."-".floor($GameNumber/200),$GameHTMLDatabaseFile);
+			If ($GameNumber == 9999){$GameDatabaseFile = $AllStarDatabaseFile;$GameNumber=0;$AllStar=True;}
 			
 			If (file_exists($GameDatabaseFile) == false){
-				echo "<title>" . $DatabaseNotFound . "</title>";
-				$GameHTML = "<h1>" . $DatabaseNotFound . "</h1>";
+				$GameBoxScore = $LeagueGeneral['OutputName']."-".$GameNumber.".php";
+				If (file_exists($GameBoxScore) == true){
+					echo "<meta http-equiv=\"refresh\" content=\"0;url=" . $LeagueOutputOption['WebsiteURL'] . "/" . $GameBoxScore . "\"/>";
+				}else{
+					echo "<title>" . $DatabaseNotFound . "</title>";
+					$GameHTML = "<h1>" . $DatabaseNotFound . "</h1>";
+				}
 			}else{
 				$Gamedb = new SQLite3($GameDatabaseFile);
 				$Query = "Select * from GameResult WHERE Number = '" . $GameNumber . "' AND Pro = '" . $TypeText . "'";
@@ -60,6 +75,12 @@ If (file_exists($DatabaseFile) == false){
 </head><body>
 <?php 
 include "Menu.php";
-if($YearH1 > 0){echo "<h1>" . $Boxscore['BoxscorefromYear'] . $YearH1 . "</h1>";}
+if($YearH1 > 0){
+	echo "<h1>" . $Boxscore['BoxscorefromYear'] . $YearH1;
+	If ($Playoff == True){echo $TopMenuLang['Playoff'];}
+	echo "</h1>";
+}elseif($AllStar == True){
+	echo "<h1>" . $TopMenuLang['AllStar'] . "</h1>";
+}
 echo($GameHTML);
 include "Footer.php";?>
